+++ src/http_async.mli
@@ -58,15 +58,15 @@ module Server : sig
   (** [run] sets up a [Tcp.Server.t] and drives the HTTP server loop with the user
       provided request-handler. *)
   val run
-    :  ?where_to_listen:Tcp.Where_to_listen.inet
-    -> ?max_connections:int
+    :  ?max_connections:int
     -> ?max_accepts_per_batch:int
     -> ?backlog:int
-    -> ?socket:([ `Unconnected ], Socket.Address.Inet.t) Socket.t
+    -> ?socket:([ `Unconnected ], ([< Socket.Address.t ] as 'a)) Socket.t
     -> ?buffer_config:Buffer_config.t
     -> ?error_handler:error_handler
+    -> where_to_listen:('a, 'b) Tcp.Where_to_listen.t
     -> (Request.t * Body.Reader.t -> (Response.t * Body.Writer.t) Deferred.t)
-    -> (Socket.Address.Inet.t, int) Tcp.Server.t Deferred.t
+    -> ('a, 'b) Tcp.Server.t Deferred.t
 
   (** [run_command] is similar to [run] but instead returns an [Async.Command.t] that can
       be used to start the async event loop from a program's entrypoint. If [interrupt] is