+++ example/echo.ml
@@ -4,7 +4,7 @@ open Http_async
 
 let () =
   Command_unix.run
-    (Server.run_command ~summary:"echo" (fun (_request, body) ->
+    (Server.run_command ~summary:"echo" (fun _addr (_request, body) ->
        let response = Response.create `Ok in
        return (response, Body.Writer.stream (Body.Reader.pipe body))))
 ;;