+++ example/hello_world.ml
@@ -4,6 +4,10 @@ open Http_async
 
 let () =
   Command_unix.run
-    (Server.run_command ~summary:"Hello world HTTP Server" (fun _request ->
+    (Server.run_command ~summary:"Hello world HTTP Server" (fun addr (request, _body) ->
+       Log.Global.info
+         "(%s): %s"
+         (Socket.Address.Inet.to_string addr)
+         (Request.path request);
        return (Response.create `Ok, Body.Writer.string "Hello World")))
 ;;